[device] Overload pick_best_physical_device
